we changed some functions

+++ router/hotel.py
@@ -4,6 +4,8 @@ from sqlalchemy.orm import Session
 from schemas import HotelBase, HotelDisplay
 from db.database import get_db
 from db import db_hotels
+from db.models import DbUser,DbHotel
+from auth.oauth2 import get_current_user
 
 router = APIRouter(
     prefix='/hotels',
@@ -11,17 +13,17 @@ router = APIRouter(
 )
 
 
-@router.post('/', response_model=HotelDisplay)
-def create_new_hotel(request: HotelBase, db: Session = Depends(get_db)):
+@router.post('/', response_model=HotelDisplay, status_code=201)
+def create_new_hotel(request: HotelBase, db: Session = Depends(get_db), current_user: DbUser = Depends(get_current_user)):
     return db_hotels.create_hotel(db, request)
 
 
-@router.get('/', response_model=List[HotelDisplay])
+@router.get('/', response_model=List[HotelDisplay], status_code=200)
 def list_hotels(db: Session = Depends(get_db)):
     return db_hotels.get_all_hotels(db)
 
 
-@router.get('/{id}', response_model=HotelDisplay)
+@router.get('/{id}', response_model=HotelDisplay, status_code=200)
 def read_hotel(id: int, db: Session = Depends(get_db)):
     hotel = db_hotels.get_hotel(db, id)
     if not hotel:
@@ -29,17 +31,23 @@ def read_hotel(id: int, db: Session = Depends(get_db)):
     return hotel
 
 
-@router.put('/{id}', response_model=HotelDisplay)
-def update_hotel(id: int, request: HotelBase, db: Session = Depends(get_db),):
-    hotel = db_hotels.update_hotel(db, id, request)
+@router.put('/{id}', response_model=HotelDisplay, status_code=200)
+def update_hotel(id: int, request: HotelBase, db: Session = Depends(get_db), current_user: DbUser = Depends(get_current_user)):
+    hotel = db.query(DbHotel).filter(DbHotel.id == id).filter(DbHotel.user_id == current_user.id).first() 
     if not hotel:
-        raise HTTPException(status_code=404, detail="Hotel not found")
-    return hotel
+        raise HTTPException(status_code=404, detail="Hotel not found or you are not authorized to delete")
+    return db_hotels.update_hotel(db, id, request)
 
 
-@router.delete('/{id}', response_model=dict)
-def delete_hotel(id: int, db: Session = Depends(get_db)):
-    hotel = db_hotels.delete_hotel(db, id)
+@router.delete('/{id}',status_code=204)
+def delete_hotel(id: int, db: Session = Depends(get_db), current_user: DbUser = Depends(get_current_user)):
+    hotel = db.query(DbHotel).filter(DbHotel.id == id).first()
     if not hotel:
-        raise HTTPException(status_code=404, detail="Hotel not found")
-    return {"detail": "Hotel deleted"}
+        raise HTTPException(status_code=404, detail="Hotel not found ")
+    if id != current_user.id:
+        raise HTTPException(status_code=403, detail="You are not authorized")
+    db_hotels.delete_hotel(db, id)
+    return 
+
+
+
